+++ messages/en/motion.php
@@ -24,6 +24,8 @@ return [
     'motion_admin_edit'             => 'Admin: edit',
     'back_start'                    => 'Back',
     'fullscreen'                    => 'Projector view (full screen)',
+    'prevnext_links_prev'           => 'Previous %TYPE%',
+    'prevnext_links_next'           => 'Next %TYPE%',
     'pdf_version'                   => 'PDF version',
     'table_caption'                 => 'Metadata of the motion',
     'table_description'             => 'The following table describes the status, the proposer and other metadata of this motion.',